requirements.txt
analytics-python==1.2.3
arrow==0.7.0
blessed==1.14.1
Django==1.9.6
django-haystack==2.5.dev1
django-picklefield==0.3.2
django-q==0.7.17
elasticsearch==2.3.0
future==0.15.2
gunicorn==19.6.0
lxml==3.6.0
parse==1.6.6
python-dateutil==2.5.3
redis==2.10.5
requests==2.10.0
six==1.10.0
urllib3==1.15.1
wcwidth==0.1.6
